Added pipeline exception tests
